Split models into specific files
